+++ rust-executor/Cargo.toml
@@ -116,6 +116,7 @@ rodio = "*"
 libc = "0.2"
 chat-gpt-lib-rs = { version = "0.5.1", git = "https://github.com/coasys/chat-gpt-lib-rs" }
 anyhow = "1.0.95"
+portpicker = "0.1.1"
 
 [dev-dependencies]
 maplit = "1.0.2"
