+++ ui/src-tauri/Cargo.toml
@@ -23,7 +23,6 @@ serde = { version = "1.0", features = ["derive"] }
 log = "0.4"
 log4rs = "1.0.0"
 nix = "0.23.1"
-portpicker = "0.1.1"
 libc = "0.2"
 directories = "4.0.1"
 opener = "0.5.0"
